Let users pass in LoaderInfo instead of DisplayObject to clear things up
Till now the SDK used a DisplayObject like a Stage to figure the URL of
the QS canvas app out. As @raichuman9 pointed out that was a stupid idea
as people could confuse and just pass in any unrelated, externally loaded
DisplayObject in which then won't come with the necressary LoaderInfo. To
fix that the SDK now takes the LoaderInfo directly so that people at least
are confronted with a much clearer decission what to pass in
(fingers crossed).